backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sniffio>=1.3.1",
    "zai-sdk>=0.2.1",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-colorama>=0.4.15",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "ALL",    # Enable ALL rules
]
ignore = [
    "D",      # pydocstyle (enable if you want docstring enforcement)
    "COM812", # conflicts with formatter
    "ISC001", # conflicts with formatter
    "T201",   # print is fine for CLI apps
    "TID252", # relative imports are fine within packages
    "FBT001", # boolean positional args are fine for simple cases
    "TC001",  # TYPE_CHECKING imports - too strict
    "TC002",  # TYPE_CHECKING imports - too strict
    "TC003",  # TYPE_CHECKING imports - too strict
    "BLE001", # blind except is sometimes needed
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = true
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.lint.pylint]
max-args = 5
max-branches = 12
max-returns = 6
max-statements = 50

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ANN",     # Don't require annotations in tests
    "PLR2004", # Allow magic values in tests
    "PLR0913", # Allow many arguments in tests (mocks)
    "FBT003",  # Allow boolean positional values in tests
    "PT011",   # Allow broad pytest.raises
    "TRY003",  # Allow long exception messages
    "EM101",   # Allow string literals in exceptions
    "SLF001",  # Allow private member access in tests
    "ARG002",  # Allow unused arguments in tests (mocks)
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
check_untyped_defs = true
no_implicit_optional = true
no_implicit_reexport = true
strict_equality = true
strict_concatenate = true
extra_checks = true

[[tool.mypy.overrides]]
module = "zai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "src.services.chat_api_service",
    "src.processors.stream_response_processor",
]
disallow_any_unimported = false

[[tool.mypy.overrides]]
module = ["src.tools.*"]
disallow_any_explicit = false
